add test case get symbol pool
